+++ Source/BenDotNet.RFID/DetectionSource.cs
@@ -6,17 +6,17 @@ namespace BenDotNet.RFID
 {
     public class DetectionSource
     {
-        public DetectionSource(ref AntennaPort antenna, ref object handle, double frequency, double rssi, double phase, DateTime time)
+        public DetectionSource(ref AntennaPort antenna, ref object handle, float frequency, float rssi, float phase, DateTime time)
         {
             this.Antenna = antenna;
             this.NotifyDetection(ref handle, frequency, rssi, phase, time);
         }
-        public void NotifyDetection(ref object handle, double frequency, double rssi, double phase, DateTime time)
+        public void NotifyDetection(ref object handle, float frequency, float rssi, float phase, DateTime time)
         {
             this.Handle = handle;
             this.NotifyDetection(frequency, rssi, phase, time);
         }
-        public void NotifyDetection(double frequency, double rssi, double phase, DateTime time)
+        public void NotifyDetection(float frequency, float rssi, float phase, DateTime time)
         {
             this.Frequency = frequency;
             this.RSSI = rssi;
@@ -26,9 +26,9 @@ namespace BenDotNet.RFID
 
         public AntennaPort Antenna { get; private set; }
         public object Handle { get; private set; }
-        public double Frequency { get; private set; }
-        public double RSSI { get; private set; }
-        public double Phase { get; private set; }
+        public float Frequency { get; private set; }
+        public float RSSI { get; private set; }
+        public float Phase { get; private set; }
         public DateTime Time { get; private set; }
     }
 }
